[login] add sections for custom env-vars, and mobile app login testing (#3)
This PR adds details about the new login field in the pilot file:
```
login:
  url: string
  username: string
  password: string
```

It also adds a section about testing login for mobile apps, emphasizing
the need
to set `login.password` so that its string literal can be handled
securely

Testing:
- inspected preview at
https://jetify-savil-login-field-in-pilot-file.mintlify.app/docs/testpilot/guides/test-authentication
- not worried about the vale linting errors below. I don't agree with
many of them. And we haven't run these rules by the rest of the docs
yet.